pyproject.toml
[project]
name = "dota2vpk"
version = "0.0.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ruamel-yaml>=0.18.15",
    "vdf>=3.4",
    "vpk>=1.4.0",
]
